:recycle: Break kebab menu out into its own component.
