+++ web/src/layouts/BaseLayout.vue
@@ -34,32 +34,7 @@
 
       <span class="text-body-2"> {{ username }} </span>
 
-      <v-menu offset-y>
-        <template #activator="{ props }">
-          <v-btn
-            icon="mdi-dots-vertical"
-            color="primary"
-            v-bind="props"
-          ></v-btn>
-        </template>
-
-        <v-list density="compact">
-          <v-list-item :to="{ name: 'StatusPage' }">
-            <template #prepend>
-              <v-icon>mdi-clock</v-icon>
-            </template>
-            <v-list-item-title class="text-body-2">
-              {{ status?.RELEASE_TAG || "loading..." }}
-            </v-list-item-title>
-          </v-list-item>
-          <v-list-item @click="logoutWrapper">
-            <template #prepend>
-              <v-icon>mdi-exit-run</v-icon>
-            </template>
-            <v-list-item-title class="text-body-2">Sign out</v-list-item-title>
-          </v-list-item>
-        </v-list>
-      </v-menu>
+      <KebabMenu />
     </template>
   </v-app-bar>
 
@@ -78,19 +53,15 @@
 </template>
 
 <script lang="ts" setup>
-import { computed, onMounted } from "vue"
-import { useAuth0 } from "@auth0/auth0-vue"
+import { computed } from "vue"
 
 import BaseBreadcrumbs from "@/components/BaseBreadcrumbs.vue"
+import KebabMenu from "@/components/base-layout/KebabMenu.vue"
 
 import { APPLICATION_NAME } from "@/config"
-import { useStatus } from "@/use/use-status"
 import { useCurrentUser } from "@/use/use-current-user"
 
-const { logout } = useAuth0()
-
 const { currentUser } = useCurrentUser()
-const { status, refresh } = useStatus()
 
 const username = computed(() => {
   if (currentUser.value === null) return "loading..."
@@ -98,17 +69,4 @@ const username = computed(() => {
   const { email } = currentUser.value
   return email.substring(0, email.indexOf("@"))
 })
-
-onMounted(async () => {
-  await refresh()
-})
-
-async function logoutWrapper() {
-  await logout({
-    logoutParams: {
-      // I would prefer to redirect to /sign-in here, but that doesn't seem to work?
-      returnTo: window.location.origin,
-    },
-  })
-}
 </script>
